Create Variables and Functions
markup the html. Create the functions for getting inputs
There will be two inputs user inputs and ai input.
Gave rock, paper, and scissor int so when we compare user choices
there won't be any need to convert string to input.

this is the initial structure design, as I refactor I will most  fixed
the logic so rather than predefined numbers, it more dynamic, might use
an array but first wanted to get the basic functionality of selecting
and comparing choices to see winner working first.