Add a function that plays a single round
The playRound function takes two parameters the user and ai choices.
Each round tracks the score if the user runs out of life, whoever
wins 3 times with each loss you lose a life. The checkScore Function
tracks the score to see if the current life is equal to the minimum life
to signal game over