Fix UI Logic
This will make the user score change when the button are clicked.

The error was occurring because the conditional statements in The
playground always evaluated to false because the character were never
lowercase in either the array or the button values. and check score
always returned a string rather than a number.